feat: 🎸 Alpha Blending and rendering distances supported
Added an alpha blending component that allows for meshes, lights and
other entities to carry an alpha value for transparency. A Quad object
in front of the vases has been added so that it demonstrates the
functionnalities of these. The lights are also ordered with these
objects and all is rendered in the correct order through sorting
entities on the fly at each render pass.